+++ common/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "autvid_common"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1"
+http = "1"
