+++ rust_stream/Cargo.toml
@@ -24,3 +24,4 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 
 # Error handling
 anyhow = "1"
+autvid_common = { path = "../common" }
